rename main to pipeline
